[OPENENGSB-876] problems with non primitives eliminated. removed fillUIDs from the interface.

+++ core/edb/src/main/java/org/openengsb/core/edb/internal/JPAObject.java
@@ -51,7 +51,7 @@ public class JPAObject {
     }
 
     public JPAObject(EDBObject o) {
-        timestamp = o.getTimestamp();
+        timestamp = o.getTimestamp().longValue();
         uid = o.getUID();
         isDeleted = o.isDeleted();
